EN-1939 porting latest frontend changes from 4.3.6.
