+++ plugins/entando-plugin-jacms/src/main/java/com/agiletec/plugins/jacms/apsadmin/global-messages_en.properties
@@ -80,6 +80,8 @@ jacms.shortcut.contentType.longDesc=Manage content types: you can edit, manage a
 menu.reload.contents=Reload the content indexes
 menu.contents.settings=Content Settings
 
+jacms.title.contentSetting.cropImage=Settings Image Crop Dimensions
+
 #Crop Editor
 cropEditor.button.close=Close
 cropEditor.button.edit=Edit
@@ -87,6 +89,7 @@ cropEditor.button.delete=Delete
 cropEditor.label.editImage=Edit Image
 cropEditor.label.edit=Edit
 cropEditor.label.delete=Delete
+cropEditor.label.flip=flip
 cropEditor.label.move=move
 cropEditor.label.crop=crop
 cropEditor.label.scale=scale
@@ -96,3 +99,5 @@ cropEditor.label.save=save
 cropEditor.label.cancel=cancel
 cropEditor.label.free=Free
 cropEditor.label.listOfImages=List of images
+cropEditor.label.cropCreated=New crop created.
+cropEditor.label.done=Done
