+++ plugins/entando-plugin-jacms/src/main/java/com/agiletec/plugins/jacms/apsadmin/global-messages_it.properties
@@ -80,6 +80,8 @@ jacms.shortcut.contentType.longDesc=Per gestire i Tipi di Contenuto: &egrave; po
 menu.reload.contents=Ricarica i contenuti
 menu.contents.settings=Impostazione contenuti
 
+jacms.title.contentSetting.cropImage=Impostazioni delle dimensioni di ritaglio delle immagini
+
 #Crop Editor
 cropEditor.button.close=chiudi
 cropEditor.button.edit=modifica
@@ -87,6 +89,7 @@ cropEditor.button.delete=cancella
 cropEditor.label.editImage=modfica immagine
 cropEditor.label.edit=Modifica
 cropEditor.label.delete=Cancella
+cropEditor.label.flip=ribalta
 cropEditor.label.move=muovi
 cropEditor.label.crop=ritaglia
 cropEditor.label.scale=scala
@@ -96,3 +99,5 @@ cropEditor.label.save=salva
 cropEditor.label.cancel=annulla
 cropEditor.label.free=libero
 cropEditor.label.listOfImages=Lista immagini
+cropEditor.label.cropCreated=Creato un nuovo ritaglio.
+cropEditor.label.done=Eseguito
