+++ plugins/entando-plugin-jacms/src/main/webapp/WEB-INF/plugins/jacms/apsadmin/jsp/common/layouts/assets-more/resource/resource.jsp
@@ -2,8 +2,8 @@
 <%@ taglib prefix="wp" uri="/aps-core" %>
 
 <link rel="stylesheet" type="text/css" href="<wp:resourceURL />administration/css/bootstrap.vertical-tabs.min.css"/>
-<link rel="stylesheet" type="text/css" href="<wp:resourceURL />administration/css/cropper.min.css"/>
-<link rel="stylesheet" type="text/css" href="<wp:resourceURL />administration/css/crop-editor.css"/>
+<link rel="stylesheet" type="text/css" href="<wp:resourceURL />administration/cropper-js/cropper.min.css"/>
+<link rel="stylesheet" type="text/css" href="<wp:resourceURL />administration/crop-editor/crop-editor.css"/>
 
 <script src="<wp:resourceURL />administration/js/jquery.sticky.js"></script>
 <script src="<wp:resourceURL />administration/js/data-tables/jquery.dataTables.min.js"></script>
@@ -37,5 +37,5 @@
     });
 </script>
 
-<script src="<wp:resourceURL />administration/js/cropper.min.js"></script>
-<script src="<wp:resourceURL />administration/js/crop-editor.js"></script>
+<script src="<wp:resourceURL />administration/cropper-js/cropper.min.js"></script>
+<script src="<wp:resourceURL />administration/crop-editor/crop-editor.js"></script>
