+++ plugins/entando-plugin-jacms/src/main/webapp/resources/administration/crop-editor/crop-editor.css
@@ -292,3 +292,11 @@
 .flex-item {
 
 }
+
+.toast-crop-editor-success {
+    margin-left: 20px;
+    position: fixed;
+    right: 20px;
+    top: 80px;
+    z-index: 1055;
+}
